Fig bug: form's data delayed
