+++ fe/src/components/languageInfo/index.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import React from 'react'
 import './style.css'
 
 const LanguageCard = ({
